refactor: add date (#27)
This PR adds a dedicated `Date` class to replace `DateTime` wherever
we're actually talking about dates. This allowed getting rid of all the
`floorToDate` calls with the hour set to `6` as a magic number.


In addition, this stores the current timezone offset for each drink. The
reason is that the `DateTime` information is lost in MongoDB. However,
this PR does not yet handle querying the drinks on a specific date
across multiple time zones (this is not easily done in MongoDB Realm, if
it's possible at all). It is currently assumed the user only drinks in
their current timezone.

The `date` stored in `DiaryEntries` (in the database) is stored in UTC.
The reason for this is that the time information, including the
timezone, is actually not important for the date.
However, this implies that when we want to compare `date` and
`startTime`, the `startTime` drink has to be converted to the timezone
it was consumed in to be comparable.